+++ py/desitarget/gaiamatch.py
@@ -68,7 +68,8 @@ gaiadatamodel = np.array([], dtype=[
 
 # ADM the current data model for READING from Gaia EDR3 files.
 inedr3datamodel = np.array([], dtype=[
-    ('SOURCE_ID', '>i8'), ('REF_CAT', 'S2'), ('RA', '>f8'), ('DEC', '>f8'),
+    ('SOURCE_ID', '>i8'), ('REF_CAT', 'S2'), ('REF_EPOCH', '>f4'), ('RA', '>f8'),
+    ('RA_ERROR', '>f8'), ('DEC', '>f8'), ('DEC_ERROR', '>f8'),
     ('PHOT_G_MEAN_MAG', '>f4'), ('PHOT_G_MEAN_FLUX_OVER_ERROR', '>f4'),
     ('PHOT_BP_MEAN_MAG', '>f4'), ('PHOT_BP_MEAN_FLUX_OVER_ERROR', '>f4'),
     ('PHOT_RP_MEAN_MAG', '>f4'), ('PHOT_RP_MEAN_FLUX_OVER_ERROR', '>f4'),
@@ -84,7 +85,8 @@ inedr3datamodel = np.array([], dtype=[
 
 # ADM the current data model for WRITING to Gaia EDR3 files.
 edr3datamodel = np.array([], dtype=[
-    ('REF_ID', '>i8'), ('REF_CAT', 'S2'), ('EDR3_RA', '>f8'), ('EDR3_DEC', '>f8'),
+    ('REF_ID', '>i8'), ('REF_CAT', 'S2'), ('REF_EPOCH', '>f4'), ('EDR3_RA', '>f8'),
+    ('EDR3_RA_IVAR', '>f8'), ('EDR3_DEC', '>f8'), ('EDR3_DEC_IVAR', '>f8'),
     ('EDR3_PHOT_G_MEAN_MAG', '>f4'), ('EDR3_PHOT_G_MEAN_FLUX_OVER_ERROR', '>f4'),
     ('EDR3_PHOT_BP_MEAN_MAG', '>f4'), ('EDR3_PHOT_BP_MEAN_FLUX_OVER_ERROR', '>f4'),
     ('EDR3_PHOT_RP_MEAN_MAG', '>f4'), ('EDR3_PHOT_RP_MEAN_FLUX_OVER_ERROR', '>f4'),
@@ -886,9 +888,10 @@ def read_gaia_file(filename, header=False, addobjid=False, dr="dr2"):
             raise ValueError(msg)
         outdata.dtype.names = edr3datamodel.dtype.names
         prefix = "EDR3"
-        # ADM the proper motion ERRORS need to be converted to IVARs.
+        # ADM the ERRORS need to be converted to IVARs.
         # ADM remember to leave 0 entries as 0.
-        for col in ['PMRA_IVAR', 'PMDEC_IVAR', 'PARALLAX_IVAR']:
+        for col in ['RA_IVAR', 'DEC_IVAR',
+                    'PMRA_IVAR', 'PMDEC_IVAR', 'PARALLAX_IVAR']:
             outcol = "{}_{}".format(prefix, col)
             w = np.where(outdata[outcol] != 0)[0]
             outdata[outcol][w] = 1./(outdata[outcol][w]**2.)